move loading to own file and small changes
